move rooter testing support to its own file
